Account for host / hostless image links
When generating an image link for a product feed, we may need to prepend
the Store's hostname. We also... may not need to! In production, `url`
returns a fully formed URL. In test, however, `url` appears to be the
local path to the image. While it's not great to write code for test, I
think this is flexible and intention-revealing enough to avoid muddying
the waters.