Re-enable default feed behavior
When we added the ability to define and manage custom Product Feeds
(based on custom ProductCatalogs), we did not preserve the old behavior
of "just give me everything". Most customers expect that behavior and do
not want to have to maintain custom catalogs. This change is a first
step back in that direction. What it does is cause the Products#index
action to return a feed of all "items" (variants) if there is no
default ProductFeed for the store.

A nice next step would be to change the routes so that the store owner
could manage and publish multiple feeds simultaneously at different
URLs. Ideally, this would allow the default behavior to continue to work.