Add code for streaming feed response
When constructing a product feed XML document for a store with a lot of
variants, processing time becomes a problem. Specifically, some services
will end a request if no data is returned within 30 seconds (looking at
you, Heroku). To ensure that our feeds are sent to clients, we switch to
a streaming response format so that data is returned almost immediately.
Most services will not end a request if data is being sent, so this
should solve the problem and scale reasonably well.